# ValueAccessor Buffer Cache