--- a/.github/workflows/docs.yml
+++ b/.github/workflows/docs.yml
@@ -40,17 +40,15 @@ jobs:
     container:
       # @note We can't update this as epydoc doesn't support python3. We'll
       # need to re-write the python docs to use sphinx
-      image: aswf/ci-openvdb:2019
+      image: aswf/ci-openvdb:2022
     steps:
     - uses: actions/checkout@v3
     - name: install_doxygen
       run: ./ci/install_doxygen.sh 1_8_11
-    - name: install_epydoc
-      run: pip install epydoc
+    # - name: install_epydoc
+    #   run: pip install epydoc
     - name: install_latex
       run: yum -y install texlive-latex-bin texlive-dvips texlive-collection-fontsrecommended texlive-collection-latexrecommended
-    - name: install_cmake
-      run: ./ci/install_cmake.sh 3.18.0
     - name: build
       run: >
         ./ci/build.sh -v
@@ -63,9 +61,9 @@ jobs:
         -DUSE_EXPLICIT_INSTANTIATION=OFF
         -DDISABLE_DEPENDENCY_VERSION_CHECKS=ON
         \"
-    - name: epydoc
-      run: |
-        epydoc --html -o /usr/local/share/doc/OpenVDB/html/python pyopenvdb
+    # - name: epydoc
+    #   run: |
+    #     epydoc --html -o /usr/local/share/doc/OpenVDB/html/python pyopenvdb
     - name: pre_deploy
       # Overwrite global SSH configuration
       #  https://github.com/peaceiris/actions-gh-pages/issues/719