… openvdb/openvdb/unittest/TestValueAccessor.cc