@jmlait 2022-10-18
Looks good, modulo the notes you have made above.

The ABI concern would be that an accessor is now longer if it comes from a new version.  So if accessors are passed through a void* we'd be in trouble.  Our requirement is that you can pass grids through void *, however, not accessors, so I think this is safe...

The grid may have a pointer to foreign accessors, but they trigger the clear() method on them, so should also be safe as they'd use the foreign vtable so respect the size.

All that said, I think I will still vote we ABI guard this new member.  We're moving to 10 anyways, and this ensures there are no unpleasant surprises.