@Idclip 2022-10-25
@jmlait @danrbailey I've updated this and added the tile fix. I've added some additional tests - we need more but this is all I have time for for now unfortunately.

Could you both give this another once over? The main thing to check now is the behaviour of `probeNode` and `probeConstNode` methods. These should now always return the desired node if it exists which is what I believe to be the desired result.

For the general case:

Before:
```c++
        if (this->isHashed(xyz)) {  // This is only true if NodeT == NodeType anyway...
            if ((std::is_same<NodeT, NodeType>::value)) {
                assert(mNode);
                return reinterpret_cast<NodeT*>(const_cast<NodeType*>(mNode));
            }
            // This can never be hit?
            return const_cast<NodeType*>(mNode)->template probeNodeAndCache<NodeT>(xyz, *mParent);
        }
        return mNext.template probeNode<NodeT>(xyz);
```

After:

```c++
        if constexpr ((std::is_same<NodeT, NodeType>::value)) {
            if (this->isHashed(xyz)) {
                assert(mNode);
                return reinterpret_cast<NodeT*>(const_cast<NodeType*>(mNode));
            }
        }

        if (NodeT::LEVEL < NodeType::LEVEL && mNode) {
            // don't need to ascend the chain, descend the tree
            return const_cast<NodeType*>(mNode)->template probeNodeAndCache<NodeT>(xyz, *mParent);
        }
        else {
            // ascend
            return mNext.template probeNode<NodeT>(xyz);
        }
```

And an example for specialized accessors:

Before, for ValueAccessor2:

```c++
        if ((std::is_same<NodeT, NodeT0>::value)) {
            if (this->isHashed0(xyz)) {
                assert(mNode0);
                return reinterpret_cast<NodeT*>(const_cast<NodeT0*>(mNode0));
            } else if (this->isHashed1(xyz)) {
                assert(mNode1);
                return const_cast<NodeT1*>(mNode1)->template probeNodeAndCache<NodeT>(xyz, *this);
            }
            return BaseT::mTree->root().template probeNodeAndCache<NodeT>(xyz, *this);
        } else if ((std::is_same<NodeT, NodeT1>::value)) {
            if (this->isHashed1(xyz)) {
                assert(mNode1);
                return reinterpret_cast<NodeT*>(const_cast<NodeT1*>(mNode1));
            }
            return BaseT::mTree->root().template probeNodeAndCache<NodeT>(xyz, *this);
        }

        // What if we are using a ValueAccessor2 on a Tree5??
        return nullptr;
```


After:

```c++
        if constexpr ((std::is_same<NodeT, NodeT0>::value)) {
            if (this->isHashed0(xyz)) {
                assert(mNode0);
                return reinterpret_cast<NodeT*>(const_cast<NodeT0*>(mNode0));
            } else if (this->isHashed1(xyz)) {
                assert(mNode1);
                return const_cast<NodeT1*>(mNode1)->template probeNodeAndCache<NodeT>(xyz, *this);
            }
            return BaseT::mTree->root().template probeNodeAndCache<NodeT>(xyz, *this);
        } else if constexpr ((std::is_same<NodeT, NodeT1>::value)) {
            if (this->isHashed1(xyz)) {
                assert(mNode1);
                return reinterpret_cast<NodeT*>(const_cast<NodeT1*>(mNode1));
            }
            return BaseT::mTree->root().template probeNodeAndCache<NodeT>(xyz, *this);
        } else if constexpr (NodeT::LEVEL < NodeT1::LEVEL) {
            // Might still be worth caching this path if a NodeT0 or NodeT1
            // are accessed in the future
            return BaseT::mTree->root().template probeNodeAndCache<NodeT>(xyz, *this);
        } else {
            // If we're accessing a node above our top most cache level then
            // there's no point trying to cache it
            return BaseT::mTree->root().template probeNode<NodeT>(xyz);
        }
```